refactor AddAppForm
reorganize code
